// src/app/models/userProfile.model.ts
export interface UserProfileModel {
    firstName: string
    lastName: string
    aadhaarNumber: string
    bloodGroup: string
    dateOfBirth: string
    email: string
    gender: string
    joiningDate: string
    phoneNumber:string
    permanentAddress: string
    presentAddress: string
    profilePic: string
    role: string
    isActive: boolean
    key: string
    isPrincipal:boolean
}